TransitionSpring API example
This is the new (mostly same) API. It's not all of the API yet; I want
to get productive on the master branch now so the actual new API will be
iterated there.

You pass down stuff to the Child spring, the same way you'd pass them
down for the dependent `Springs`. You can also optionally carry around a
`data` field for misc info regarding the spring, e.g. title. The reason
we do that is because otherwise, you can't access that piece of data
from the final list of items, since it might not be there anymore (the
current item's already gone, but still animating out in our
interpolating values).

But this `data` field reveals a (solvable) problem in this new API
(same problem exists in the dependent Springs API): if you destructure
it (line 141), and then apply `rest` to `Child`'s `to`, the whole thing
won't animate, because we count on mutating the config and reading them
from Child (mutating only for the perf optimization ofc). Splatting onto
`rest` is shallow cloning the whole thing so the pointer to the mutating
`config` is lost.

There are a few solutions.

- Place `data` somewhere else, so that the user has no reason to
  destructure `config`. This only pushes the problem further.

- Make the config object opaque (does not conflict with the first
  point). This way you can't destructure. I dislike unnecessarily opaque
  objects, but here it's not strictly "unnecessary".

- Let children `raf` (make it so for the dependent Spring too). You keep
  passing a non-mutative config (allows destructuring ofc), and the
  child now keeps the state of the velocity and all. This has a few
  implications:

  * It was probably the necessary thing to do in the first place. We
    can't have an owner `raf`-ing independently and then the children
    `raf`-ing too and expect them to keep the same clock and everything
    interpolating smoothly.

  * Children will probably need a handler to an owner's `forceUpdate` so
    that it can call it when it finishes raf-ing for the owner to force
    update and pass down the new interpolated value that doesn't have
    that child anymore.

  * For dependent `Springs`, you'd need to pass down all the configs to
    each Child, plus a cursor to indicate which config you want that
    specific child to be; you can't calculate a child's value without
    the values it's depending on. The API might get weird.